+++ zeitgitter/stamper.py
@@ -20,7 +20,9 @@
 
 # Timestamp creation
 
+import logging as _logging
 import os
+import sys
 import re
 import threading
 import time
@@ -29,6 +31,77 @@ from pathlib import Path
 import gnupg
 import zeitgitter.commit
 import zeitgitter.config
+
+
+logging = _logging.getLogger('stamper')
+
+
+def get_nick(domain):
+    fields = domain.split('.')
+    for i in fields:
+        if i != '' and i != 'igitt' and i != 'zeitgitter' and 'stamp' not in i:
+            return i
+    sys.exit("Please specify a keyid")
+
+
+def create_key(gpg, keyid):
+    name, mail = keyid.split(' <')
+    mail = mail[:-1]
+    gki = gpg.gen_key_input(name_real=name, name_email=mail,
+            key_type='eddsa', key_curve='Ed25519', key_usage='sign')
+    # Force to not ask (non-existant) user for passphrase
+    gki = '%no-protection\n' + gki
+    key = gpg.gen_key(gki)
+    if key.fingerprint is None:
+        system.exit("Cannot create PGP key for %s: %s" % (keyid, key.stderr))
+    logging.info("Created PGP key %s" % key.fingerprint)
+    return key.fingerprint
+
+
+def get_keyid(keyid, domain, gnupg_home):
+    """Find (and possibly create) a key ID by trying the following, in order:
+    1. `keyid` is not `None`:
+       1. `keyid` matches exactly one secret key: return the hexadecimal
+           key ID of that key.
+       2. `keyid` does not match any existing key, but matches the pattern
+           "Name <email>": Create an ed25519 key and return its ID.
+       3. Else fail.
+    2. `keyid` is `None`:
+       1. If there is exactly one secret key in the keyring, return that one.
+       2. If there are no secret keys in the keyring, create an ed25519 key
+          named "Nickname Timestamping Service <domain-to-email>", where
+          "Nickname" is the capitalized first dot-separated part of `domain`
+          which is not "zeitgitter", "igitt", or "*stamp*"; and
+          "domain-to-email" is `domain` with its first dot is replaced by an @.
+       3. Else fail.
+    The key is only created if the next start would chose that key
+    (this obviates the need to monkey-patch the config file)."""
+    gpg = gnupg.GPG(gnupghome=gnupg_home)
+    if keyid is not None:
+        keyinfo = gpg.list_keys(secret=True, keys=keyid)
+        if len(keyinfo) == 1:
+            return keyinfo[0]['keyid']
+        elif re.match("^[A-Z][A-Za-z0-9 ]+ <[-_a-z0-9.]+@[-a-z0-9.]+>$",
+                keyid) and len(keyinfo) == 0:
+            return create_key(gpg, keyid)
+        else:
+            if len(keyinfo) > 0:
+                sys.exit("Too many secret keys matching key '%s'" % keyid)
+            else:
+                sys.exit("No secret keys match keyid '%s', "
+                    "use the form 'Description <email>' if zeitgitterd "
+                    "should create one automatically with that name" % keyid)
+    else:
+        keyinfo = gpg.list_keys(secret=True)
+        if len(keyinfo) == 1:
+            return keyinfo[0]['keyid']
+        elif len(keyinfo) == 0:
+            nick = get_nick(domain)
+            maildomain = domain.replace('.', '@', 1)
+            return create_key(gpg, "%s Timestamping Service <%s>"
+                    % (nick.capitalize(), maildomain))
+        else:
+            sys.exit("Please specify a keyid in the configuration file")
 
 
 class Stamper:
@@ -40,7 +113,7 @@ class Stamper:
         self.url = zeitgitter.config.arg.own_url
         self.keyid = zeitgitter.config.arg.keyid
         self.gpgs = [gnupg.GPG(gnupghome=zeitgitter.config.arg.gnupg_home)]
-        self.keyinfo = self.gpg().list_keys(keys=self.keyid)
+        self.keyinfo = self.gpg().list_keys(True, keys=self.keyid)
         if len(self.keyinfo) == 0:
             raise ValueError("No keys found")
         self.fullid = self.keyinfo[0]['uids'][0]
@@ -61,7 +134,6 @@ class Stamper:
                     home.symlink_to(zeitgitter.config.arg.gnupg_home)
                 nextgpg = gnupg.GPG(gnupghome=home.as_posix())
                 self.gpgs.append(nextgpg)
-                print(nextgpg)
                 return nextgpg
             else:
                 # Rotate list left and return element wrapped around (if the list
@@ -69,7 +141,6 @@ class Stamper:
                 nextgpg = self.gpgs[0]
                 self.gpgs = self.gpgs[1:]
                 self.gpgs.append(nextgpg)
-                print(nextgpg)
                 return nextgpg
 
     def sig_time(self):
@@ -80,12 +151,14 @@ class Stamper:
         return self.pubkey
 
     def valid_tag(self, tag):
-        """Tag validity defined in doc/Protocol.md"""
+        """Tag validity defined in doc/Protocol.md
+
+        Switching to `pygit2.reference_is_valid_name()` should be considered
+        when this function is widely availabe in installations
+        (was only added on 2018-10-17)"""
         # '$' always matches '\n' as well. Don't want this here.
-        if '\n' in tag:
-            return False
-        return (re.match('^[a-z][-._a-z0-9]{,99}$', tag, re.IGNORECASE)
-                and ".." not in tag)
+        return (re.match('^[_a-z][-._a-z0-9]{,99}$', tag, re.IGNORECASE)
+                and ".." not in tag and not '\n' in tag)
 
     def valid_commit(self, commit):
         # '$' always matches '\n' as well. Don't want this here.